feat: share BaseAudioContext interface through macros + remove generator for audio contexts
