+++ src/audio_context.rs
@@ -1,22 +1,3 @@
-// -------------------------------------------------------------------------- //
-// -------------------------------------------------------------------------- //
-//                                                                            //
-//                                                                            //
-//                                                                            //
-//    ██╗    ██╗ █████╗ ██████╗ ███╗   ██╗██╗███╗   ██╗ ██████╗               //
-//    ██║    ██║██╔══██╗██╔══██╗████╗  ██║██║████╗  ██║██╔════╝               //
-//    ██║ █╗ ██║███████║██████╔╝██╔██╗ ██║██║██╔██╗ ██║██║  ███╗              //
-//    ██║███╗██║██╔══██║██╔══██╗██║╚██╗██║██║██║╚██╗██║██║   ██║              //
-//    ╚███╔███╔╝██║  ██║██║  ██║██║ ╚████║██║██║ ╚████║╚██████╔╝              //
-//     ╚══╝╚══╝ ╚═╝  ╚═╝╚═╝  ╚═╝╚═╝  ╚═══╝╚═╝╚═╝  ╚═══╝ ╚═════╝               //
-//                                                                            //
-//                                                                            //
-//    - This file has been generated ---------------------------              //
-//                                                                            //
-//                                                                            //
-// -------------------------------------------------------------------------- //
-// -------------------------------------------------------------------------- //
-
 use std::collections::HashMap;
 use std::io::Cursor;
 use std::sync::{Arc, Mutex};
@@ -49,29 +30,19 @@ pub(crate) struct NapiAudioContext {
 
 impl NapiAudioContext {
     pub fn create_js_class(env: &Env) -> Result<JsFunction> {
-        env.define_class(
-            "AudioContext",
-            constructor,
-            &[
-                Property::new("currentTime")?.with_getter(get_current_time),
-                Property::new("sampleRate")?.with_getter(get_sample_rate),
-                Property::new("listener")?.with_getter(get_listener),
-                Property::new("state")?.with_getter(get_state),
-                Property::new("decodeAudioData")?.with_method(decode_audio_data),
-                // ----------------------------------------------------
-                // Methods and attributes specific to AudioContext
-                // ----------------------------------------------------
-                Property::new("baseLatency")?.with_getter(get_base_latency),
-                Property::new("outputLatency")?.with_getter(get_output_latency),
-                Property::new("sinkId")?.with_getter(get_sink_id),
-                Property::new("setSinkId")?.with_method(set_sink_id),
-                Property::new("resume")?.with_method(resume),
-                Property::new("suspend")?.with_method(suspend),
-                Property::new("close")?.with_method(close),
-                // private
-                Property::new("__initEventTarget__")?.with_method(init_event_target),
-            ],
-        )
+        let interface = base_audio_context_interface![
+            Property::new("baseLatency")?.with_getter(get_base_latency),
+            Property::new("outputLatency")?.with_getter(get_output_latency),
+            Property::new("sinkId")?.with_getter(get_sink_id),
+            Property::new("setSinkId")?.with_method(set_sink_id),
+            Property::new("resume")?.with_method(resume),
+            Property::new("suspend")?.with_method(suspend),
+            Property::new("close")?.with_method(close),
+            // [non spec] Bind with JS EventTarget
+            Property::new("__initEventTarget__")?.with_method(init_event_target)
+        ];
+
+        env.define_class("AudioContext", constructor, &interface)
     }
 
     pub fn unwrap(&self) -> &AudioContext {
@@ -89,10 +60,6 @@ impl NapiAudioContext {
     // We need to clean things around so that the js object can be garbage collected.
     // But we also need to wait so that the previous tsfn.call is executed.
     // This is not clean, but don't see how to implement that properly right now.
-
-    // no ended events for nodes that are created by OfflineAudioContext
-    // remove directive once implemented
-    #[allow(dead_code)]
     pub fn clear_thread_safe_listener(&self, store_id: String) {
         std::thread::sleep(std::time::Duration::from_millis(1));
         let mut tsfn_store = self.tsfn_store.lock().unwrap();
@@ -174,19 +141,12 @@ fn constructor(ctx: CallContext) -> Result<JsUndefined> {
     };
     ctx.env.wrap(&mut js_this, napi_audio_context)?;
 
-    js_this.define_properties(&[
-        // this must be put on the instance and not in the prototype to be reachable
-        Property::new("Symbol.toStringTag")?
-            .with_value(&ctx.env.create_string("AudioContext")?)
-            .with_property_attributes(PropertyAttributes::Static),
-    ])?;
-
-    // test symbol as property name
-    // let test_symbol = ctx.env.symbol_for("test").unwrap();
-    // js_this.set_property(test_symbol, &ctx.env.create_string("test").unwrap())?;
+    js_this.define_properties(&[Property::new("Symbol.toStringTag")?
+        .with_value(&ctx.env.create_string("AudioContext")?)
+        .with_property_attributes(PropertyAttributes::Static)])?;
 
     // -------------------------------------------------
-    // Bind AudioDestination
+    // Bind AudioDestination - requires Symbol.toStringTag
     // -------------------------------------------------
     let store_ref: &mut napi::Ref<()> = ctx.env.get_instance_data()?.unwrap();
     let store: JsObject = ctx.env.get_reference_value(store_ref)?;
@@ -197,143 +157,7 @@ fn constructor(ctx: CallContext) -> Result<JsUndefined> {
     ctx.env.get_undefined()
 }
 
-#[js_function]
-fn get_current_time(ctx: CallContext) -> Result<JsNumber> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let obj = napi_obj.unwrap();
-
-    let current_time = obj.current_time();
-    ctx.env.create_double(current_time)
-}
-
-#[js_function]
-fn get_sample_rate(ctx: CallContext) -> Result<JsNumber> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let obj = napi_obj.unwrap();
-
-    let sample_rate = obj.sample_rate() as f64;
-    ctx.env.create_double(sample_rate)
-}
-
-// use a getter so we can lazily create the listener on first call and retrieve it afterward
-#[js_function]
-fn get_listener(ctx: CallContext) -> Result<JsObject> {
-    let mut js_this = ctx.this_unchecked::<JsObject>();
-
-    // reproduce lazy instanciation strategy from rust crate
-    if js_this.has_named_property("__listener__").ok().unwrap() {
-        js_this.get_named_property("__listener__")
-    } else {
-        let store_ref: &mut napi::Ref<()> = ctx.env.get_instance_data()?.unwrap();
-        let store: JsObject = ctx.env.get_reference_value(store_ref)?;
-        let ctor: JsFunction = store.get_named_property("AudioListener")?;
-        let js_obj = ctor.new_instance(&[&js_this])?;
-        js_this.set_named_property("__listener__", &js_obj)?;
-
-        Ok(js_obj)
-    }
-}
-
-#[js_function]
-fn get_state(ctx: CallContext) -> Result<JsString> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let obj = napi_obj.unwrap();
-
-    let state = obj.state();
-    let state_str = match state {
-        AudioContextState::Suspended => "suspended",
-        AudioContextState::Running => "running",
-        AudioContextState::Closed => "closed",
-    };
-
-    ctx.env.create_string(state_str)
-}
-
-// ----------------------------------------------------
-// METHODS
-// ----------------------------------------------------
-
-// @todo - async version
-#[js_function(1)]
-fn decode_audio_data(ctx: CallContext) -> Result<JsObject> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let context = napi_obj.unwrap();
-
-    let js_buffer = ctx.get::<JsArrayBuffer>(0)?.into_value()?;
-    let cursor = Cursor::new(js_buffer.to_vec());
-    let audio_buffer = context.decode_audio_data_sync(cursor);
-
-    match audio_buffer {
-        Ok(audio_buffer) => {
-            // create js audio buffer instance
-            let store_ref: &mut napi::Ref<()> = ctx.env.get_instance_data()?.unwrap();
-            let store: JsObject = ctx.env.get_reference_value(store_ref)?;
-            let ctor: JsFunction = store.get_named_property("AudioBuffer")?;
-            let mut options = ctx.env.create_object()?;
-            options.set("__internal_caller__", ctx.env.get_null())?;
-
-            // populate with audio buffer
-            let js_audio_buffer = ctor.new_instance(&[options])?;
-            let napi_audio_buffer = ctx.env.unwrap::<NapiAudioBuffer>(&js_audio_buffer)?;
-            napi_audio_buffer.populate(audio_buffer);
-
-            Ok(js_audio_buffer)
-        }
-        Err(e) => Err(napi::Error::from_reason(e.to_string())),
-    }
-}
-
-// ----------------------------------------------------
-// Methods and attributes specific to AudioContext
-// ----------------------------------------------------
-#[js_function]
-fn resume(ctx: CallContext) -> Result<JsObject> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let clone = Arc::clone(&napi_obj.context);
-
-    ctx.env.execute_tokio_future(
-        async move {
-            clone.resume().await;
-            Ok(())
-        },
-        |&mut env, _val| env.get_undefined(),
-    )
-}
-
-#[js_function]
-fn suspend(ctx: CallContext) -> Result<JsObject> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let clone = Arc::clone(&napi_obj.context);
-
-    ctx.env.execute_tokio_future(
-        async move {
-            clone.suspend().await;
-            Ok(())
-        },
-        |&mut env, _val| env.get_undefined(),
-    )
-}
-
-#[js_function]
-fn close(ctx: CallContext) -> Result<JsObject> {
-    let js_this = ctx.this_unchecked::<JsObject>();
-    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
-    let clone = Arc::clone(&napi_obj.context);
-
-    ctx.env.execute_tokio_future(
-        async move {
-            clone.close().await;
-            Ok(())
-        },
-        |&mut env, _val| env.get_undefined(),
-    )
-}
+base_audio_context_impl!(NapiAudioContext);
 
 #[js_function]
 fn get_base_latency(ctx: CallContext) -> Result<JsNumber> {
@@ -383,9 +207,54 @@ fn set_sink_id(ctx: CallContext) -> Result<JsUndefined> {
     ctx.env.get_undefined()
 }
 
-// ----------------------------------------------------
-// Private Event Target initialization
-// ----------------------------------------------------
+#[js_function]
+fn resume(ctx: CallContext) -> Result<JsObject> {
+    let js_this = ctx.this_unchecked::<JsObject>();
+    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
+    let clone = Arc::clone(&napi_obj.context);
+
+    ctx.env.execute_tokio_future(
+        async move {
+            clone.resume().await;
+            Ok(())
+        },
+        |&mut env, _val| env.get_undefined(),
+    )
+}
+
+#[js_function]
+fn suspend(ctx: CallContext) -> Result<JsObject> {
+    let js_this = ctx.this_unchecked::<JsObject>();
+    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
+    let clone = Arc::clone(&napi_obj.context);
+
+    ctx.env.execute_tokio_future(
+        async move {
+            clone.suspend().await;
+            Ok(())
+        },
+        |&mut env, _val| env.get_undefined(),
+    )
+}
+
+#[js_function]
+fn close(ctx: CallContext) -> Result<JsObject> {
+    let js_this = ctx.this_unchecked::<JsObject>();
+    let napi_obj = ctx.env.unwrap::<NapiAudioContext>(&js_this)?;
+    let clone = Arc::clone(&napi_obj.context);
+
+    ctx.env.execute_tokio_future(
+        async move {
+            clone.close().await;
+            Ok(())
+        },
+        |&mut env, _val| env.get_undefined(),
+    )
+}
+
+// ----------------------------------------------------------
+// [non spec] Bind with JS EventTarget
+// ----------------------------------------------------------
 #[js_function]
 fn init_event_target(ctx: CallContext) -> Result<JsUndefined> {
     let js_this = ctx.this_unchecked::<JsObject>();
